refactor(console): refactor `useSubscriptionPlans` hook
