+++ packages/console/src/pages/TenantSettings/components/Skeleton/index.tsx
